remove blocking-async-wrapped node REST Swift API
in favor of pure async;
