+++ bindings/ergo-lib-c-core/src/rest/api.rs
@@ -2,6 +2,5 @@
 
 pub mod callback;
 pub mod node;
-pub mod node_async;
 pub mod request_handle;
 pub mod runtime;
