+++ bindings/ergo-lib-c/src/rest/api.rs
@@ -1,6 +1,5 @@
 //! REST API for the services in Ergo ecosystem (node, explorer, etc.)
 
 pub mod node;
-pub mod node_async;
 pub mod request_handle;
 pub mod runtime;
