+++ bindings/ergo-lib-c/src/rest/api/node.rs
@@ -1,7 +1,8 @@
+use ergo_lib_c_core::rest::api::callback::CompletionCallback;
 use ergo_lib_c_core::rest::api::node::rest_api_node_get_info;
+use ergo_lib_c_core::rest::api::request_handle::RequestHandlePtr;
 use ergo_lib_c_core::rest::api::runtime::RestApiRuntimePtr;
 use ergo_lib_c_core::rest::node_conf::NodeConfPtr;
-use ergo_lib_c_core::rest::node_info::NodeInfoPtr;
 use ergo_lib_c_core::Error;
 use ergo_lib_c_core::ErrorPtr;
 
@@ -9,8 +10,9 @@ use ergo_lib_c_core::ErrorPtr;
 pub unsafe extern "C" fn ergo_lib_rest_api_node_get_info(
     runtime_ptr: RestApiRuntimePtr,
     node_conf_ptr: NodeConfPtr,
-    node_info_out: *mut NodeInfoPtr,
+    callback: CompletionCallback,
+    request_handle_out: *mut RequestHandlePtr,
 ) -> ErrorPtr {
-    let res = rest_api_node_get_info(runtime_ptr, node_conf_ptr, node_info_out);
+    let res = rest_api_node_get_info(runtime_ptr, node_conf_ptr, callback, request_handle_out);
     Error::c_api_from(res)
 }
