--- bindings/ergo-lib-ios/Sources/ErgoLib/RestNodeApiAsync.swift
@@ -1,37 +1,0 @@
-import Foundation
-import ErgoLibC
-
-
-enum RestNodeApiError: Error {
-    case misc(String)
-}
-
-class RestNodeApiAsync {
-    internal var pointer: RestApiRuntimePtr
-    
-    /// Create ergo ``RestNodeApiAsync`` instance 
-    init() throws {
-        var ptr: RestApiRuntimePtr?
-        let error = ergo_lib_rest_api_runtime_create(&ptr)
-        try checkError(error)
-        self.pointer = ptr!
-    }
-    
-    /// GET on /info endpoint
-    func getInfo(
-        nodeConf: NodeConf,
-        closure: @escaping (Result<NodeInfo, Error>) -> Void
-    ) throws -> RequestHandle {
-        
-        let completion = wrapClosure(closure)
-        var requestHandlerPtr: RequestHandlePtr?
-        let error = ergo_lib_rest_api_node_get_info_async(self.pointer, nodeConf.pointer, 
-            completion, &requestHandlerPtr)
-        try checkError(error)
-        return RequestHandle(withRawPtr: requestHandlerPtr!)
-    }
-    
-    deinit {
-        ergo_lib_rest_api_runtime_delete(self.pointer)
-    }
-}
