+++ bindings/ergo-lib-ios/Tests/ErgoLibTests/RestNodeApiTests.swift
@@ -6,10 +6,30 @@ import XCTest
 final class RestNodeApiTests: XCTestCase {
 
     func testGetInfo() throws {
+        let expectation = self.expectation(description: "getInfo")
         let nodeConf = try NodeConf(withAddrString: "213.239.193.208:9053")
         let restNodeApi = try RestNodeApi()
-        let nodeInfo = try restNodeApi.getInfo(nodeConf: nodeConf)
-        XCTAssert(!nodeInfo.getName().isEmpty)
+        let _ = try restNodeApi.getInfo(nodeConf: nodeConf,
+            closure: { (res: Result<NodeInfo, Error>) -> () in 
+                switch res {
+                    case .success(let nodeInfo): 
+                        XCTAssert(!nodeInfo.getName().isEmpty)
+                    case .failure(let error): 
+                        XCTFail(error.localizedDescription)
+                }
+                expectation.fulfill()
+            })
+        waitForExpectations(timeout: 5, handler: nil)
+    }
+
+    func testGetInfoAbort() throws {
+        let nodeConf = try NodeConf(withAddrString: "213.239.193.208:9053")
+        let restNodeApi = try RestNodeApi()
+        let handle = try restNodeApi.getInfo(nodeConf: nodeConf,
+            closure: { (res: Result<NodeInfo, Error>) -> () in 
+                XCTFail("this should not be called")
+            })
+        handle.abort()
     }
 
     // need macOS 12.0 on Github GA
